Modularise into different classes
